+++ config/index.js
@@ -24,14 +24,24 @@ config.set({
       },
       dashboard: {
         title: 'Groundwire Trading Preferences',
-        prefs: [
-          {name: 'account', title: 'Robinhood Account', active: false, icon: '\uF2BE'},
-          {name: 'risk', title: 'Risk Settings', active: false, icon: '\uF2C8'},
-          {name: 'chase', title: 'Chase Coefficient', active: false, icon: '\uF201'},
-          {name: 'screener', title: 'Stock Picks', active: false, icon: '\uF03A'},
-          {name: 'strategy', title: 'Stop Loss Strategy', active: false, icon: '\uF0AB'},
-          {name: 'activate', title: 'Activate', active: false, icon: '\uF011'}
-        ]
+        prefs: {
+          account: {name: 'account', title: 'Robinhood Account', active: false, icon: '\uF2BE', description: 'Connect or disconnect your robinhood account'},
+          risk: {name: 'risk', title: 'Risk Settings', active: false, icon: '\uF2C8', description: 'Manage the risk settings for your automated trading alogorithm'},
+          chase: {name: 'chase', title: 'Chase Coefficient', active: false, icon: '\uF201', description: 'Set the chase speed of the stop loss algorithm'},
+          screener: {name: 'screener', title: 'Stock Picks', active: false, icon: '\uF03A', description: 'Set the criteria for the MSN screener stock picking feature'},
+          strategy: {name: 'strategy', title: 'Stop Loss Strategy', active: false, icon: '\uF0AB', description: 'Select a specific trailing stop loss algorithm'},
+          activate: {name: 'activate', title: 'Activate', active: false, icon: '\uF011', description: 'Turn on or off the automated trader for actively trading on your Robinhood account'}
+        }
+      }
+    },
+    ui: {
+      button: {
+        save: {
+          title: 'Save'
+        },
+        saved: {
+          title: 'Saved'
+        }
       }
     }
   },
@@ -90,6 +100,16 @@ config.set({
       url: 'https://api.groundwire.co/v1/strategy',
       method: 'GET',
       options: {}
+    },
+    preferences: {
+      url: 'https://api.groundwire.co/v1/preferences',
+      method: 'GET',
+      options: {}  
+    },
+    savePreferences: {
+      url: 'https://api.groundwire.co/v1/preferences',
+      method: 'POST',
+      options: {}  
     }
   }
 });