+++ src/components/forms/activate.vue
@@ -1,0 +1,3 @@
+<template>
+  <h1>Activate vue</h1>
+</template>