+++ src/components/forms/index.js
@@ -1,0 +1,15 @@
+import AccountForm from 'components/forms/account';
+import RiskForm from 'components/forms/risk';
+import ScreenerForm from 'components/forms/screener';
+import StrategyForm from 'components/forms/strategy';
+import ActivateForm from 'components/forms/activate';
+import ChaseForm from 'components/forms/chase';
+
+export default {
+  AccountForm,
+  RiskForm,
+  ScreenerForm,
+  StrategyForm,
+  ActivateForm,
+  ChaseForm
+}