+++ src/components/forms/screener.vue
@@ -1,0 +1,3 @@
+<template>
+  <h1>Screener vue</h1>
+</template>