+++ src/state/actions.js
@@ -1,4 +1,0 @@
-export const toggleAuthentication = ({ dispatch }) => {
-	alert("called");
-  dispatch('TOGGLE_AUTHENTICATION')
-}