+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.10.0] - 2017-11-24
+### Added
+- Setting strategy and saving to the db
+- Vuex actions dispatched from components
+
 ## [0.9.0] - 2017-11-11
 ### Added
 - Generic switch control